+++ test/bigtest/tests/package-usage-consolidation-test.js
@@ -92,7 +92,8 @@ describe('PackageShowUsageConsolidation', () => {
           await PackageShowPage.usageConsolidation.content.whenLoaded();
         });
 
-        describe('waiting for axe to run', () => {
+        // TODO: Try to run test after STCOM-788
+        describe.skip('waiting for axe to run', () => {
           beforeEach(async () => {
             await PackageShowPage.whenLoaded();
             a11yResults = await axe.run();
@@ -272,6 +273,10 @@ describe('PackageShowUsageConsolidation', () => {
 
     it('should show Summary table', () => {
       expect(PackageShowPage.usageConsolidation.content.summaryTable.isPresent).to.be.true;
+    });
+
+    it('should show four columns in Summary table', () => {
+      expect(PackageShowPage.usageConsolidation.content.summaryTable.columnCount).to.be.equal(4);
     });
 
     it('should show Usage in correct format', () => {