+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * Clear Usage Consolidation data when leaving Package/Title/Resource View page. (UIEH-1013)
 * Package Show page | Export Titles. (UIEH-946)
 * Display Usage data metric selection for Total usage on Summary table. (UIEH-994)
+* Remove Actions button dropdown from Title/Resource Summary Table. (UIEH-1014)
 
 ## [5.0.0] (https://github.com/folio-org/ui-eholdings/tree/v5.0.0) (2020-10-15)